add context in countDonw
