+++ src/Pages/Home/Components/CountDown/index.tsx
@@ -1,6 +1,53 @@
+import { differenceInSeconds } from "date-fns";
+import { useContext, useEffect, useState } from "react";
+import { CyclesContext } from "../../Home";
 import { CountdownContainer, Separator } from "./styles";
 
 export function CountDown() {
+  const { activeCycle, activeCycleId, markCurrentAsFinished } =
+    useContext(CyclesContext);
+
+  const [amountSecondsPassed, setamountSecondsPassed] = useState(0);
+  const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0;
+
+  const currentSeconds = activeCycle ? totalSeconds - amountSecondsPassed : 0;
+
+  const minutesAmount = Math.floor(currentSeconds / 60);
+  const secondsAmount = currentSeconds % 60;
+
+  const minutes = String(minutesAmount).padStart(2, "0");
+  const seconds = String(secondsAmount).padStart(2, "0");
+
+  useEffect(() => {
+    if (activeCycle) {
+      document.title = `${minutes}: ${seconds}`;
+    }
+  }, [minutes, seconds, activeCycle]);
+
+  useEffect(() => {
+    let interval: number;
+
+    if (activeCycle) {
+      interval = setInterval(() => {
+        const secondsDifefrence = differenceInSeconds(
+          new Date(),
+          activeCycle.startDate
+        );
+
+        if (secondsDifefrence >= totalSeconds) {
+          markCurrentAsFinished();
+          setamountSecondsPassed(totalSeconds);
+          clearInterval(interval);
+        } else {
+          setamountSecondsPassed(secondsDifefrence);
+        }
+      }, 1000);
+    }
+    return () => {
+      clearInterval(interval);
+    };
+  }, [activeCycle, totalSeconds, activeCycleId, markCurrentAsFinished]);
+
   return (
     <CountdownContainer>
       <span>{minutes[0]}</span>
